Working addition
addition working both for ints and reals

+++ src/main/java/pl/ziemniakoss/simplecompiler/VariableType.java
@@ -6,6 +6,6 @@ public enum VariableType {
 
 	@Override
 	public String toString() {
-		return this == INTEGER ? "i32" : "float";
+		return this == INTEGER ? "i32" : "double";
 	}
 }
